docs: add examples, documentation site, and dev planning; clean tooling
- Add tutorial examples (notebooks + utils) for GARCH, multivariate,
  regime-switching, threshold/STAR, risk, diagnostics, and workflows
- Add MkDocs documentation site (API, user guide, theory, benchmarks)
- Add dev/ sprint planning docs and Read the Docs config
- Untrack generated artifacts already in .gitignore (mutants/,
  .hypothesis/, logs/, __pycache__, .coverage, .mutmut-cache)
- Fix pyright strict errors in archbox/ (float64 dtype casts,
  None-narrowing) and configure ruff per-file-ignores for examples/
- Make check-yaml --unsafe for mkdocs.yml custom tags

+++ archbox/models/aparch.py
@@ -123,7 +123,7 @@ class APARCH(VolatilityModel):
         gammas = np.full(self.q, 0.0)
         betas = np.full(self.p, 0.90)
         delta = np.array([2.0])
-        return np.concatenate([[omega], alphas, gammas, betas, delta])
+        return np.concatenate([[omega], alphas, gammas, betas, delta], dtype=np.float64)
 
     @property
     def param_names(self) -> list[str]:
